% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{applyDeadModel}
\alias{applyDeadModel}
\title{Applies death risk model to new cohort}
\usage{
applyDeadModel(connectionDetails, cdmDatabaseSchema, cohortDatabaseSchema,
  oracleTempSchema = NULL, cohortTable, cohortId)
}
\arguments{
\item{connectionDetails}{The connection details to connect to a database}

\item{cdmDatabaseSchema}{The common data model schema}

\item{cohortDatabaseSchema}{The database schema containing the cohort you wish to apply the death risk prediction to}

\item{oracleTempSchema}{The temp schema for oracle (default is NULL)}

\item{cohortTable}{The table containing the cohort you want to apply the death risk model to}

\item{cohortId}{The cohort_definition_id defining the cohort you wish to apply the death risk model to}
}
\value{
A dataframe with each subjectId and cohortStartDate in the cohort with the predicted death risk as the column 'value'
}
\description{
Applies death risk model to new cohort
}
\details{
This will return a predicted risk of being dead at the cohort_start_date for each subject_id in the cohort
}
