% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{viewDeadCoefficients}
\alias{viewDeadCoefficients}
\title{View DEAD prediction model coefficients}
\usage{
viewDeadCoefficients(packageName = "DeadModel")
}
\description{
View DEAD prediction model coefficients
}
\details{
This will show a table with the covaraite name and coefficient value
}
