% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validateDeadModel}
\alias{validateDeadModel}
\title{Validate death plp prediction model}
\usage{
validateDeadModel(connectionDetails, cdmDatabaseSchema, cohortDatabaseSchema,
  oracleTempSchema = NULL, cohortTable, targetId, outcomeId,
  packageName = "DeadModel")
}
\arguments{
\item{oracleTempSchema}{The temp schema for oracle (default is NULL)}

\item{cohortTable}{The name of the cohort table}

\item{targetId}{The cohort definition id of the target population}

\item{outcomeId}{The cohort definition id of the outcome}

\item{connectioDetails}{The connections details for connecting to the CDM}

\item{cdmDatabaseschema}{The schema holding the CDM data}

\item{cohortDatabaseschema}{The schema holding the cohort table}
}
\value{
A list with the performance and plots
}
\description{
Validate death plp prediction model
}
\details{
This will run and evaluate an exisitng death plpModel
}
